--- src/routes/test/+page.svelte
@@ -1,24 +1,0 @@
-<script lang="ts">
-    let cssClasses: string = ""
-    let compiledCss: string = ""
-
-    const onChange = async (e: Event) => {
-        compiledCss = (
-            await fetch(`/test?css=${cssClasses}`).then(r => r.json())
-        ).output
-        console.log(compiledCss)
-    }
-</script>
-
-<svelte:head>
-    {@html `<style type="text/css">${compiledCss}</style>`}
-</svelte:head>
-
-<input type="text" 
-    placeholder="class names here" 
-    on:input={(e) => onChange(e)}
-    bind:value={cssClasses}/>
-
-<div id="css-target" class={cssClasses}>
-    Style me!
-</div>
