+++ tailwind.config.ts
@@ -4,11 +4,9 @@ import { skeleton } from '@skeletonlabs/tw-plugin'
 import typography from '@tailwindcss/typography'
 
 const config = {
-	// 2. Opt for dark mode to be handled via the class method
 	darkMode: 'class',
 	content: [
 		'./src/**/*.{html,js,svelte,ts}',
-		// 3. Append the path to the Skeleton package
 		join(require.resolve(
 			'@skeletonlabs/skeleton'),
 			'../**/*.{html,js,svelte,ts}'
@@ -18,7 +16,6 @@ const config = {
 		extend: {},
 	},
 	plugins: [
-		// 4. Append the Skeleton plugin (after other plugins)
 		skeleton({
 			themes: { preset: [ "skeleton" ] }
 		}),
